+++ CHANGELOG.md
@@ -14,9 +14,8 @@
  - Change RaisedButton to ElevatedButton in timer example
  - Overhaul the draggables api to fix issues relating to local vs global positions
  - Preventing errors caused by the premature use of size property on game
- - Add `PositionComponent.position` constructor
- - Add `SpriteComponent.position` constructor
- - Add `SpriteAnimationComponent.position` constructor
+ - `PositionComponent` remove abstract
+ - Component constrctor unified
 
 ## 1.0.0-rc6
  - Use `Offset` type directly in `JoystickAction.update` calculations